@@ -14,7 +14,7 @@
     <description>easy-retry-client-starter</description>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
     </properties>
 
     <dependencies>
