@@ -2,7 +2,7 @@ package com.aizuda.easy.retry.client.model;
 
 import lombok.Data;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 /**
  * 生成idempotentId模型
