@@ -3,9 +3,9 @@ package com.aizuda.easy.retry.server.web.model.request;
 import com.aizuda.easy.retry.server.web.model.base.BaseQueryVO;
 import lombok.Data;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Pattern;
 
 
 /**
