@@ -2,10 +2,10 @@ package com.aizuda.easy.retry.server.web.model.request;
 
 import com.aizuda.easy.retry.common.core.enums.RetryStatusEnum;
 import lombok.Data;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
 
-import javax.validation.constraints.Pattern;
+import jakarta.validation.constraints.Pattern;
 import java.util.List;
 
 /**
